Split out stuff for embed linux distro
